requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
shapely==2.0.3
openmeteo-requests
requests-cache 
retry-requests 
pandas 




